#3157: Add falcon7b demo | add demo test | debug hang
