+++ models/demos/falcon7b/demo/input_data.json
@@ -1,0 +1,97 @@
+[
+    {
+"question": "write a short poem about London in English"
+            },
+         {
+"question": "write a short poem about Madrid in Spanish"
+            },
+         {
+"question": "write a short poem about Madrid in English"
+            },
+         {
+"question": "write a short poem about Paris in English"
+            },
+         {
+"question": "write a short poem about Paris in French"
+            },
+         {
+"question": "write a short poem about Istanbul in English"
+            },
+         {
+"question": "write a short poem about Shanghai in English"
+            },
+         {
+"question": "write a short poem about Lagos in English"
+            },
+         {
+"question": "what is the capital of USA? "
+            },
+         {
+"question": "what is the capital of Canada? "
+            },
+         {
+"question": "what is the capital of UK? "
+            },
+         {
+"question": "what is the capital of Germany? "
+            },
+         {
+"question": "what is the capital of France? "
+            },
+         {
+"question": "what is the capital of Japan? "
+            },
+         {
+"question": "what is the capital of India? "
+            },
+         {
+"question": "what is the capital of China? "
+            },
+         {
+"question": "what is the currency of Cuba? "
+            },
+         {
+"question": "what is the currency of Lebanon? "
+            },
+         {
+"question": "what is the currency of Brazil? "
+            },
+         {
+"question": "what is the currency of Australia? "
+            },
+         {
+"question": "what is the currency of Jamaica? "
+            },
+         {
+"question": "what is the currency of Egypt? "
+            },
+         {
+"question": "what is the currency of Uzbekistan? "
+            },
+         {
+"question": "what is the currency of Argentina? "
+            },
+         {
+"question": "describe the geographic location of London in UK"
+            },
+         {
+"question": "describe the geographic location of Toronto in Canada"
+            },
+         {
+"question": "describe the geographic location of Madrid in Spain"
+            },
+         {
+"question": "describe the geographic location of Paris in France"
+            },
+         {
+"question": "describe the geographic location of Rome in Italy"
+            },
+         {
+"question": "describe the geographic location of Istanbul in Turkey"
+            },
+         {
+"question": "describe the geographic location of Shanghai in China"
+            },
+         {
+"question": "describe the geographic location of Lagos in Nigeria"}
+        ]
