+++ models/demos/falcon7b/reference/cpu_demo.py
