+++ models/demos/falcon7b/tests/test_falcon_decoder.py
@@ -7,11 +7,11 @@ import pytest
 from loguru import logger
 
 import tt_lib
-from models.falcon7b.reference.hf_modeling_falcon import (
+from models.demos.falcon7b.reference.hf_modeling_falcon import (
     FalconForCausalLM,
 )
-from models.falcon7b.tt.falcon_attention import TtFalconAttention
-from models.falcon7b.model_config import (
+from models.demos.falcon7b.tt.falcon_decoder import TtFalconDecoderLayer
+from models.demos.falcon7b.tt.model_config import (
     get_model_config,
     get_tt_cache_path,
 )
@@ -22,16 +22,16 @@ from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import (
 from models.utility_functions import torch2tt_tensor, tt2torch_tensor
 
 
-class PytorchFalconAttentionModel(torch.nn.Module):
+class PytorchFalconDecoderModel(torch.nn.Module):
     def __init__(self, hf_reference_model, layer_num):
         super().__init__()
-        self.attention = hf_reference_model.transformer.h[layer_num].self_attention
+        self.decoder = hf_reference_model.transformer.h[layer_num]
 
         # Disable dropout
-        self.attention.eval()
+        self.decoder.eval()
 
     def forward(self, x, alibi, attention_mask, layer_past, use_cache):
-        result = self.attention(
+        result = self.decoder(
             hidden_states=x,
             alibi=alibi,
             attention_mask=attention_mask,
@@ -41,13 +41,14 @@ class PytorchFalconAttentionModel(torch.nn.Module):
         return result
 
 
-def run_test_FalconAttention_inference(
+def run_test_FalconDecoder_inference(
     device,
     model_version,
     llm_mode,
     batch,
     seq_len,
     kv_cache_len,
+    layer_num,
     pcc,
     model_config,
     tt_cache_path,
@@ -59,15 +60,15 @@ def run_test_FalconAttention_inference(
     hugging_face_reference_model.eval()
     configuration = hugging_face_reference_model.config
     state_dict = hugging_face_reference_model.state_dict()
-    use_cache = True
-    user_id = 0
 
-    # Prepare input
+    # Prepare input ========================================================================
     torch.manual_seed(0)
-    layer_num = 0
+    decoder_input = (torch.rand(batch, seq_len, configuration.hidden_size) * 2) - 1
     base_url = "transformer.h"
     max_position_embeddings = 2048
     head_dim = configuration.hidden_size // configuration.n_head
+    use_cache = True
+    user_id = 0
 
     # Generate input, attention_mask, and kv_cache --------------------------------------
     # TODO: Generate attention_mask on device
@@ -77,13 +78,13 @@ def run_test_FalconAttention_inference(
         assert q_len % 32 == 0, "For prefill, seq_len must be multiple of 32!"
         assert kv_cache_len == 0, "For prefill, no kv_cache is passed in!"
 
-        attention_input = (torch.rand(batch, q_len, configuration.hidden_size) * 2) - 1
+        decoder_input = (torch.rand(batch, q_len, configuration.hidden_size) * 2) - 1
         attention_mask_bool = torch.ones(batch, 1, q_len, kv_len, dtype=bool).triu(
             diagonal=1
         )
         layer_past = None
 
-        tt_attention_input = torch2tt_tensor(attention_input.unsqueeze(1), device)
+        tt_decoder_input = torch2tt_tensor(decoder_input.unsqueeze(1), device)
         tt_attention_mask = torch2tt_tensor(
             (attention_mask_bool * -100000).expand(-1, configuration.n_head, -1, -1),
             device,
@@ -99,16 +100,15 @@ def run_test_FalconAttention_inference(
         assert batch % 32 == 0, "For decode, batch must be multiple of 32!"
         assert q_len == 1, "For decode, q_len must be 1!"
 
-        attention_input = (torch.rand(batch, q_len, configuration.hidden_size) * 2) - 1
-        # attention_input = (torch.rand(batch, q_len, 4544) * 2) - 1
+        decoder_input = (torch.rand(batch, q_len, configuration.hidden_size) * 2) - 1
         attention_mask_bool = torch.zeros(batch, 1, q_len, kv_len, dtype=bool)
         attention_mask_bool[:, :, :, -1] = True
         k_cache = torch.rand(batch, kv_cache_len, head_dim)
         v_cache = torch.rand(batch, kv_cache_len, head_dim)
         layer_past = (k_cache, v_cache)
 
-        tt_attention_input = torch2tt_tensor(
-            attention_input.unsqueeze(1).transpose(0, 2), device
+        tt_decoder_input = torch2tt_tensor(
+            decoder_input.unsqueeze(1).transpose(0, 2), device
         )
 
         kv_len_padded = (kv_len + 31) // 32 * 32
@@ -121,11 +121,7 @@ def run_test_FalconAttention_inference(
         )
         tt_attention_mask = torch2tt_tensor(
             (attention_mask_bool_padded.transpose(0, 2) * -100000).expand(
-                -1,
-                configuration.n_head,
-                -1,
-                -1
-                # -1, 71, -1, -1
+                -1, configuration.n_head, -1, -1
             ),
             device,
         )
@@ -142,50 +138,46 @@ def run_test_FalconAttention_inference(
             f"Llm mode {llm_mode} is not supported! Must be one of prefill or decode."
         )
 
-    # PyTorch output --------------------------------------------------------------------
-    pytorch_FalconAttention_model = PytorchFalconAttentionModel(
+    # PyTorch output =======================================================================
+    pytorch_FalconDecoder_model = PytorchFalconDecoderModel(
         hugging_face_reference_model, layer_num
     )
-    pytorch_out, pytorch_layer_present = pytorch_FalconAttention_model(
-        attention_input,
+    pytorch_out, pytorch_layer_present = pytorch_FalconDecoder_model(
+        x=decoder_input,
         alibi=None,
         attention_mask=attention_mask_bool,
         layer_past=layer_past,
         use_cache=use_cache,
     )
 
-    # TT hardware execution -------------------------------------------------------------
-    tt_FalconAttention_model = TtFalconAttention(
+    # TT hardware execution =================================================================
+    tt_FalconDecoder_model = TtFalconDecoderLayer(
         device,
         state_dict,
-        # None,
         base_url,
         layer_num,
-        configuration.hidden_size,
-        configuration.n_head,
-        # 4544,
-        # 71,
+        configuration,
         max_position_embeddings,
         model_config,
         tt_cache_path,
     )
 
-    tt_out, tt_layer_present = tt_FalconAttention_model(
-        tt_attention_input,
+    tt_out, tt_layer_present = tt_FalconDecoder_model(
+        hidden_states=tt_decoder_input,
+        llm_mode=llm_mode,
         alibi=None,
         attention_mask=tt_attention_mask,
-        llm_mode=llm_mode,
         user_id=user_id,
         layer_past=tt_layer_past,
         layer_past_len=kv_cache_len,
         use_cache=use_cache,
     )
     tt_out = tt2torch_tensor(tt_out).squeeze(1)
+
     tt_layer_present = (
         tt2torch_tensor(tt_layer_present[0]).squeeze(1),
         tt2torch_tensor(tt_layer_present[1]).squeeze(1),
     )
-
     if llm_mode == "decode":
         tt_out = tt_out.transpose(0, 1)
     tt_layer_present = (
@@ -212,9 +204,9 @@ def run_test_FalconAttention_inference(
     does_pass = does_pass and does_pass2
 
     if does_pass:
-        logger.info("Falcon Attention output Passed!")
+        logger.info("Falcon Decoder output Passed!")
     else:
-        logger.warning("Falcon Attention output Failed!")
+        logger.warning("Falcon Decoder output Failed!")
         assert does_pass, f"PCC value is lower than {pcc}"
 
 
@@ -227,16 +219,17 @@ def run_test_FalconAttention_inference(
     ids=["prefill_seq128", "decode_batch32"],
 )
 @pytest.mark.parametrize(
-    "model_version, pcc",
-    (("tiiuae/falcon-7b-instruct", 0.98),),
+    "model_version, layer_num, pcc",
+    (("tiiuae/falcon-7b-instruct", 0, 0.98),),
 )
 @pytest.mark.parametrize("model_config_str", ("BFLOAT16-DRAM", "BFLOAT16-L1"))
-def test_FalconAttention_inference(
+def test_FalconDecoder_inference(
     model_version,
     llm_mode,
     batch,
     seq_len,
     kv_cache_len,
+    layer_num,
     pcc,
     model_config_str,
     model_location_generator,
@@ -245,13 +238,14 @@ def test_FalconAttention_inference(
     model_config = get_model_config(model_config_str)
     tt_cache_path = get_tt_cache_path(model_version)
 
-    run_test_FalconAttention_inference(
+    run_test_FalconDecoder_inference(
         device,
         model_version,
         llm_mode,
         batch,
         seq_len,
         kv_cache_len,
+        layer_num,
         pcc,
         model_config,
         tt_cache_path,
