+++ models/demos/falcon7b/tt/falcon_mlp.py
